requirements.txt
PyAutoGUI~=0.9.54
pillow~=10.2.0
mimesis~=13.1.0
selenium~=4.17.2
opencv-python~=4.9.0.80
numpy~=1.26.3
pyperclip~=1.8.2